requirements.txt
agno==2.0.9
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
docstring-parser==0.17.0
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.4
groq==0.31.1
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.1
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.43
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
